update metrics page to pull from local metrics
 * Replace the extant metrics pages with the alternative version that
   pulls from es6.
