+++ admin/static/js/metrics/metrics.es6.js
@@ -8,7 +8,7 @@ var c3 = require('c3/c3.js');
 var keen = require('keen-js');
 var keenDataviz = require('keen-dataviz');
 var keenAnalysis = require('keen-analysis');
-
+var $ = require('jquery');
 
 var client = new keenAnalysis({
     projectId: keenProjectId,
@@ -32,6 +32,301 @@ var yearColor = '#0CEB92';
 var dayColor = '#0C93F5';
 var privateColor = '#F20066';
 var publicColor = '#FFD300';
+
+
+function MetricParams(params) {
+    this.params = params;
+
+    // params
+    //   eventCollection -- name of metrics api endpoint
+    //   targetProperty -- property to extract in response
+    //   timeframe -- <str> or <obj> with .start & .end keys
+    //     <str> - previous_${count}_${period}
+    //       $count -- int >= 1
+    //       $period -- days, weeks, months
+    //   timezone -- timezone (do we use this?)
+    //   interval -- used by some respReformatters (interval for responses)
+    //   groupBy -- <list> properties to group by ???
+    //   respReformatter -- what type of response processor is this
+    //     singleval
+    //       targetProperty
+    //     interval
+    //       interval
+    //       targetProperty
+    //     extract-list
+    //       listKey
+    //       nameKey
+    //       resultKey
+    //     just-return
+    //       *no properties*
+    //     grouped-interval
+    //       interval
+    //       groupBy[0], groupBy[1]
+    //       targetProperty
+    //     table
+    //       tableProperties
+    //   listKey -- (used by extract-list) --- ???
+    //   nameKey -- (used by extract-list) --- ???
+    //   resultKey -- (used by extract-list) --- ???
+
+    //   endpoint -- either "reports" or "query", defaults to "reports"
+
+    //   filters -- used for NodeLogs & legacy download counts
+    //     property_name
+    //     operator
+    //     property_value
+    //     timezone
+}
+
+function makeMetricsUrl(newQuery) {
+    var queryParams = {};
+
+    if (newQuery.timeframe) {
+        if (typeof newQuery.timeframe === 'object') {
+            queryParams["timeframeStart"] = newQuery.timeframe.start;
+            queryParams["timeframeEnd"] = newQuery.timeframe.end;
+        }
+        else {
+            queryParams["timeframe"] = newQuery.timeframe;
+        }
+    }
+
+    var urlRoot;
+    if (newQuery.endpoint && newQuery.endpoint === 'query') {
+        urlRoot = metricsUrl.replace('reports', 'query') + newQuery.eventCollection + '/';
+    }
+    else { // default to 'reports' endpoint
+        urlRoot = metricsUrl + newQuery.eventCollection + '/recent/';
+    }
+    var newUrl = new URL(urlRoot);
+    for (var key in queryParams) {
+        if (queryParams.hasOwnProperty(key)) {
+            newUrl.searchParams.append(key, queryParams[key]);
+        }
+    }
+
+    return newUrl;
+}
+
+// take a property name like 'foo.bar' and a data structure and return data.foo.bar
+function _diveIntoProperties(propertyName, data) {
+    var propertydive = propertyName.split('.');
+    var tmp = data;
+    propertydive.forEach(property => {
+        if (tmp !== undefined) {
+            tmp = tmp[property];
+        }
+    });
+    return tmp;
+}
+
+
+function _query2Promise(queryParams) {
+    // Localize timezone if none is set
+    if (!queryParams.timezone) {
+        queryParams.timezone = new Date().getTimezoneOffset() * -60;
+    }
+
+    return $.get(makeMetricsUrl(queryParams));
+}
+
+/**
+ * Take a list of MetricParams objects: call them, resolve them, turn them into outputs
+ *
+ * @method _resolveQueries
+ * @param {Array} queries - Array of MetricParams objects to call and resolve.
+ * @return {Array} array of response outputs
+ * @return {Object} a single response output
+ */
+function _resolveQueries(queries) {
+
+    var promises = [];
+    queries.forEach(query => {
+        promises.push(_query2Promise(query.params));
+    });
+
+    // if multiple promises, return array of outputs
+    // if single promise, return output w/o array wrapper
+    return promises.length > 1 ? Promise.all(promises) : promises[0];
+}
+
+/**
+ * Take a MetricParams object and a metric api response and remap the response to a structure
+ * suitable for passing to the `.chart()` method of a keenDataviz object.
+ *
+ * @method reformatResponse
+ * @param {MetricParams} newQuery - MetricParams object representing a query to the metrics endpoint
+ * @param {Object} res - response from metrics api endpoint
+ * @return {Object} - remapped results data structure
+ */
+function reformatResponse(newQuery, res) {
+    var keenFormatResp = {
+        query: newQuery,
+    };
+
+    if (newQuery.respReformatter === "interval") {
+        // take list of data and remap each to:
+        //   {value: $targetProperty, timeframe: {start: $start, end: $end}}
+        // currently assumes that both input and output are bucketed on daily intervals
+        keenFormatResp['result'] = [];
+        var interval = 1;
+        if (newQuery.interval !== 'daily') {
+            console.debug('ResponseReformatter: Unsupported interval in query:', newQuery);
+        }
+
+        res.data.forEach(entry => {
+            var timestamp = entry['attributes']['report_date'];
+            var remade = {
+                timeframe: getInterval(timestamp, interval),
+                value: entry['attributes'][newQuery.targetProperty],
+            };
+            keenFormatResp['result'].push(remade);
+        });
+    }
+    else if (newQuery.respReformatter === "grouped-interval") {
+        // take list of data and group by property by interval (currently only daily interval
+        // is supported
+        // newQuery.groupBy is a two-entry array
+        //   groupBy[0] is the name of the group label in the output data
+        //   groupBy[1] is the name of the group label in the input data (from res)
+        //   targetProperty is the name of the property (must be dereferenced)
+        var interval = 1;
+        if (newQuery.interval !== "daily") {
+            console.debug('ResponseReformatter: Unsupported interval in query:', newQuery);
+        }
+
+        var groupKeySrc = newQuery.groupBy[0];
+        var groupKeyTgt = newQuery.groupBy[1];
+        var intervals = {};
+        res.data.forEach(entry => {
+            var reportDate = entry.attributes.report_date;
+            if (!intervals[reportDate]) {
+                intervals[reportDate] = {
+                    timeframe: getInterval(reportDate, interval),
+                    value: [],
+                };
+            }
+            intervals[reportDate].value.push({
+                groupKeySrc: entry.attributes[groupKeyTgt],
+                result: _diveIntoProperties(newQuery.targetProperty, entry.attributes),
+            });
+        });
+
+        keenFormatResp['result'] = Object.values(intervals);
+        keenFormatResp['result'].forEach(record => {
+            record.value.sort((a, b) => a.groupKeySrc.localeCompare(b.groupKeySrc));
+        });
+    }
+    else if (newQuery.respReformatter === "singleval") {
+        // output should be a single value in a key called "result"
+        // input may be an object or a one-element array; unwrap array
+        //   targetProperty is the name of the property in the unwrapped response
+        var thisData;
+        if (Array.isArray(res.data)) {
+            if (res.data.length > 1) {
+                console.debug('In "singleval" respReformatter, expected 1 data point, but ' +
+                              'received ' + res.data.length + '.', res.data);
+            }
+            if (res.data.length >= 1) {
+                thisData = res.data[0];
+            }
+        }
+        else {
+            thisData = res.data;
+        }
+
+        if (thisData && thisData.attributes) {
+            keenFormatResp['result'] = _diveIntoProperties(newQuery.targetProperty, thisData.attributes);
+        }
+        else {
+            keenFormatResp['result'] = 'No data';
+        }
+    }
+    else if (newQuery.respReformatter === "table") {
+        // turn input data into list of objects
+        //  tableProperties is a whitelist of properties to transfer to the output list
+        keenFormatResp['result'] = [];
+        res.data.forEach(thisData => {
+            var remade = {};
+            newQuery.tableProperties.forEach(propName => {
+                remade[propName] = _diveIntoProperties(propName, thisData.attributes);
+            });
+            keenFormatResp['result'].push(remade);
+        });
+    }
+    else if (newQuery.respReformatter === "just-return") {
+        keenFormatResp['result'] = [];
+        res.data.forEach(thisData => {
+            var attrs = thisData['attributes'];
+            var interval = 1;
+            var remade = {
+                timeframe: getInterval(attrs.report_date, interval),
+                attributes: attrs,
+            };
+            keenFormatResp['result'].push(remade);
+        });
+    }
+    else if (newQuery.respReformatter === "return-sublist") {
+        keenFormatResp['result'] = [];
+        res.data.attributes[newQuery.listKey].forEach(thisData => {
+            var interval = 1;
+            var remade = {
+                timeframe: getInterval(thisData.date, interval),
+                value: thisData[newQuery.targetProperty],
+            };
+            keenFormatResp['result'].push(remade);
+        });
+    }
+    else if (newQuery.respReformatter === "sum-over") {
+        keenFormatResp['result'] = 0;
+        res.data.forEach(thisData => {
+            var attrs = thisData['attributes'];
+            keenFormatResp['result'] += _diveIntoProperties(newQuery.targetProperty, attrs);
+        });
+    }
+    else if (newQuery.respReformatter === "extract-list") {
+        // take a list of daily tallies and turn them into a different list
+        // input data format (res.data): [
+        //   {attributes: {report_date: "2022-12-03", $listKey: [
+        //     {$nameKey: "foo", $resultKey: 3}, {$nameKey: "bar", $resultKey: 10}
+        //   ]}},
+        //   {attributes: {report_date: "2022-12-04", $listKey: [
+        //     {$nameKey: "foo", $resultKey: 5}, {$nameKey: "bar", $resultKey: 12}
+        //   ]}},
+        // ]
+        // output data format (keenFormatResp.result): [
+        //   {timeframe: {start: $start, end: $end}, value: [
+        //     {key: "foo", result: 3}, {key: "bar", result: 10}
+        //   ]},
+        //   {timeframe: {start: $start, end: $end}, value: [
+        //     {key: "foo", result: 5}, {key: "bar", result: 12}
+        //   ]},
+        // ]
+        keenFormatResp['result'] = [];
+        var interval = 1;
+
+        res.data.forEach(thisData => {
+            var attrs = thisData['attributes'];
+            var remade = {
+                timeframe: getInterval(attrs.report_date, interval),
+                value: [],
+            };
+            attrs[newQuery.listKey].forEach(entry => {
+                remade.value.push({
+                    key: entry[newQuery.nameKey],
+                    result: _diveIntoProperties(newQuery.resultKey, entry),
+                });
+            });
+            keenFormatResp['result'].push(remade);
+        });
+    }
+    else {
+        console.debug('ResponseReformatter: reformatter not supported?', newQuery.respReformatter);
+    }
+
+    return keenFormatResp;
+}
+
 
 /**
  * Configure a timeframe for the past day for a keen query
@@ -59,34 +354,34 @@ var getOneDayTimeframe = function(daysBack, monthsBack) {
     today.setDate(today.getDate() + 1);
     end = today.toISOString();
     return {
-        "start": start,
-        "end": end
+        "start": start.replace('T00:00:00.000Z', ''),
+        "end": end.replace('T00:00:00.000Z', ''),
     };
 };
 
+
 /**
- * Configure a time frame for a day x days ago (end) and y days prior to x (start)
+ * Configure a timeframe ending on endDate (end) and starting priorDays prior (start)
  *
- * @method getVariableDayTimeframe
- * @param {Integer} endDaysBack - the number of days back to set as the end day
- * @param {Integer} totalDays - the number of days back to reach the start day
+ * @method getInterval
+ * @param {str} endDate - the last day of the timeframe (dt format: mm-dd-yyyy)
+ * @param {Integer} priorDays - the number of days back to reach the start day
  * @return {Object} the keen-formatted timeframe
  */
-var getVariableDayTimeframe = function(endDaysBack, totalDays) {
+var getInterval = function(endDate, priorDays) {
     var start = null;
     var end = null;
-    var date = new Date();
 
-    date.setUTCDate(date.getDate() - endDaysBack);
-    date.setUTCHours(0, 0, 0, 0, 0);
-
+    var date = new Date(endDate);
     end = date.toISOString();
 
-    date.setDate(date.getDate() - totalDays);
+    date = new Date(date.getTime() - (priorDays * 24 * 60 * 60 * 1000));
+    date.setUTCHours(0, 0, 0, 0, 0);
     start = date.toISOString();
+
     return {
-        "start": start,
-        "end": end
+        "start": start.replace('T00:00:00.000Z', ''),
+        "end": end.replace('T00:00:00.000Z', ''),
     };
 };
 
@@ -97,11 +392,13 @@ var getVariableDayTimeframe = function(endDaysBack, totalDays) {
  * @param {Object} metric - metric result object to get the date from
  * @return {String} the title for the monthly metric chart
  */
+// KRA-NOTE - reach into metric here
 var getMetricTitle = function(metric, type) {
 
     if (metric.params.timeframe.start) {
 
-        var monthNames = ["January", "February", "March", "April", "May", "June",
+        var monthNames = [
+            "January", "February", "March", "April", "May", "June",
             "July", "August", "September", "October", "November", "December"
         ];
         var date = null;
@@ -111,162 +408,22 @@ var getMetricTitle = function(metric, type) {
         if (type === "month") {
             date = new Date(metric.params.timeframe.start);
             title =  monthNames[date.getUTCMonth()] + " to " + monthNames[(date.getUTCMonth() + 1)%12];
-        } else if (type === "day") {
+        }
+        else if (type === "day") {
             date = metric.params.timeframe.start.replace('T00:00:00.000Z', '');
             end = metric.params.timeframe.end.replace('T00:00:00.000Z', '');
             title =  date + " until " + end;
         }
 
-    } else {
+    }
+    else {
         title = metric.params.timeframe;
     }
 
     return title;
 };
 
-
-var differenceGrowthBetweenMetrics = function(metric1, metric2, totalMetric, element, colors) {
-    var percentOne;
-    var percentTwo;
-    var differenceMetric = new keenDataviz()
-        .el(element)
-        .chartType("metric")
-        .chartOptions({
-            suffix: '%'
-        })
-        .title(' ')
-        .height(defaultHeight)
-        .prepare();
-
-    if (colors) {
-        differenceMetric.colors([colors]);
-    }
-
-    client.run([
-        metric1,
-        metric2,
-        totalMetric
-    ]).then(function(res) {
-        var metricOneResult = res[0].result;
-        var metricTwoResult = res[1].result;
-        var totalResult = res[2].result;
-
-        percentOne = (metricOneResult/totalResult)*100;
-        percentTwo = (metricTwoResult/totalResult)*100;
-
-        var data = {
-            "result": percentOne - percentTwo
-        };
-
-        differenceMetric.parseRawData(data).render();
-    }).catch(function(err) {
-        differenceMetric.message(err.message);
-    });
-};
-
-
-var renderPublicPrivatePercent = function(publicMetric, privateMetric, element) {
-    var result;
-    var differenceMetric = new keenDataviz()
-        .el(element)
-        .type("pie")
-        .title(' ')
-        .prepare();
-
-    client.run([
-        publicMetric,
-        privateMetric,
-    ], function(err, res) {
-
-        result = [
-            {'result': res[0].result, 'label': 'public'}, {'result': res[1].result, 'label': 'private'}
-        ];
-
-        var data = {
-            "result": result
-        };
-
-        differenceMetric.parseRawData(data).render();
-    });
-};
-
-
-var renderCalculationBetweenTwoQueries = function(query1, query2, element, differenceType, calculationType, colors) {
-    var result;
-    var differenceMetric;
-
-    if (calculationType === "percentage") {
-        differenceMetric = new keenDataviz()
-            .el(element)
-            .type("metric")
-            .title(' ')
-            .height(defaultHeight)
-            .chartOptions({
-                suffix: '%'
-            })
-            .prepare();
-    } else {
-        differenceMetric = new keenDataviz()
-            .el(element)
-            .height(defaultHeight)
-            .chartType("metric")
-            .title(' ')
-            .prepare();
-    }
-
-    if (colors) {
-        differenceMetric.colors([colors]);
-    }
-
-    differenceMetric.title(getMetricTitle(query1, differenceType));
-
-    client.run([
-        query1,
-        query2
-    ]).then(function(res) {
-        var metricOneResult = res[0].result;
-        var metricTwoResult = res[1].result;
-        if (calculationType === "subtraction") {
-            result = metricOneResult - metricTwoResult;
-        } else if (calculationType === "percentage") {
-            result = (metricOneResult/metricTwoResult) * 100;
-        } else if (calculationType === "division") {
-            result = metricOneResult/metricTwoResult;
-        }
-
-        var data = {
-            "result": result
-        };
-
-        differenceMetric.parseRawData(data).render();
-    })
-    .catch(function(err){
-        differenceMetric.message(err.message);
-    });
-};
-
-var getWeeklyUserGain = function() {
-    var queries = [];
-    var timeframes = [];
-
-    // get timeframes from 1 day back through 7 days back for the full week
-    for (var i = 1; i < 8; i++) {
-        var timeframe = getOneDayTimeframe(i, null);
-        var query = new keenAnalysis.Query("sum", {
-            eventCollection: "user_summary",
-            targetProperty: "status.active",
-            timeframe: timeframe
-        });
-        queries.push(query);
-        timeframes.push(timeframe);
-    }
-
-    return {"queries": queries, "timeframes": timeframes};
-
-};
-
 var renderKeenMetric = function(element, type, query, height, colors, keenClient) {
-
     if (!keenClient) {
         keenClient = client;
     }
@@ -327,88 +484,187 @@ var renderNodeLogsForOneUserChart = function(user_id) {
         });
 };
 
+// called from pageview collection
+var differenceGrowthBetweenMetrics = function(query1, query2, totalQuery, element, colors) {
+
+    var percentOne;
+    var percentTwo;
+
+    var differenceMetric = new keenDataviz()
+        .el(element)
+        .chartType("metric")
+        .chartOptions({
+            suffix: '%'
+        })
+        .title(' ')
+        .height(defaultHeight)
+        .prepare();
+
+    if (colors) {
+        differenceMetric.colors([colors]);
+    }
+
+    _resolveQueries([
+        query1,
+        query2,
+        totalQuery,
+    ]).then(function(res) {
+        var queryOneResult = reformatResponse(query1.params, res[0]).result;
+        var queryTwoResult = reformatResponse(query2.params, res[1]).result;
+        var totalResult    = reformatResponse(totalQuery.params, res[2]).result;
+
+        percentOne = (queryOneResult/totalResult)*100;
+        percentTwo = (queryTwoResult/totalResult)*100;
+
+        var data = {
+            "result": percentOne - percentTwo
+        };
+
+        differenceMetric.parseRawData(data).render();
+    }).catch(function(err) {
+        differenceMetric.message(err.message);
+    });
+};
+
+var renderCalculationBetweenTwoQueries = function(query1, query2, element, differenceType,
+                                                  calculationType, colors) {
+    var result;
+    var differenceMetric;
+
+    if (calculationType === "percentage") {
+        differenceMetric = new keenDataviz()
+            .el(element)
+            .type("metric")
+            .title(' ')
+            .height(defaultHeight)
+            .chartOptions({
+                suffix: '%'
+            })
+            .prepare();
+    }
+    else {
+        differenceMetric = new keenDataviz()
+            .el(element)
+            .height(defaultHeight)
+            .chartType("metric")
+            .title(' ')
+            .prepare();
+    }
+
+    if (colors) {
+        differenceMetric.colors([colors]);
+    }
+
+    differenceMetric.title(getMetricTitle(query1, differenceType));
+
+    _resolveQueries([
+        query1,
+        query2,
+    ]).then(function(res) {
+        var metricOneResult = reformatResponse(query1.params, res[0]).result;
+        var metricTwoResult = reformatResponse(query2.params, res[1]).result;
+
+        if (calculationType === "subtraction") {
+            result = metricOneResult - metricTwoResult;
+        }
+        else if (calculationType === "percentage") {
+            result = (metricOneResult/metricTwoResult) * 100;
+        }
+        else if (calculationType === "division") {
+            result = metricOneResult/metricTwoResult;
+        }
+
+        var data = {
+            "result": result
+        };
+
+        differenceMetric.parseRawData(data).render();
+    })
+    .catch(function(err){
+        differenceMetric.message(err.message);
+    });
+};
+
+var renderMetricsForInsts = function(instQuery, eachMetric) {
+    _resolveQueries([instQuery]).then(res => {
+        eachMetric.forEach(me => {
+            var newParams = $.extend({tableProperties: ['institution_name',  me[1]]}, instQuery.params);
+            var newRes = reformatResponse(newParams, res);
+            var chart = new keenDataviz()
+                .el(me[0])
+                .height(institutionTableHeight)
+                .title(' ')
+                .type('table')
+                .prepare();
+
+            var metricChart = chart.data(newRes);
+            metricChart.dataset.sortRows("desc", function(row) {
+                return row[1];
+            });
+            metricChart.render();
+        });
+    }).fail(function (jqXHR, textStatus, err) {
+        eachMetric.forEach(me => {
+            var chart = new keenDataviz()
+                .el(me[0])
+                .height(institutionTableHeight)
+                .title(' ')
+                .type('table')
+                .prepare();
+            chart.message(err.message);
+        });
+    });
+};
+
+var renderMetric = function(element, type, query, height, colors) {
+
+    var chart = new keenDataviz()
+        .el(element)
+        .height(height)
+        .title(' ')
+        .type(type)
+        .prepare();
+
+    if (colors) {
+        chart.colors([colors]);
+    }
+
+    _resolveQueries([query]).then(res => {
+        var newRes = reformatResponse(query.params, res);
+        if (newRes['result'] === 'No data') {
+            chart.message('No data');
+            return;
+        }
+        var metricChart = chart.data(newRes);
+        metricChart.dataset.sortRows("desc", function(row) {
+            return row[1];
+        });
+        metricChart.render();
+    }).fail(function (jqXHR, textStatus, err) {
+        chart.message(err.message);
+    });
+};
+
 
 // Common Queries
 
 // Active user count! - Total Confirmed Users of the OSF
-var activeUsersQuery = new keenAnalysis.Query("sum", {
-    event_collection: "user_summary",
-    target_property: "status.active",
+var activeUsersQuery = new MetricParams({
+    eventCollection: "user_summary",
+    targetProperty: "active",
     timeframe: "previous_1_days",
-    timezone: "UTC"
+    timezone: "UTC",
+    respReformatter: "singleval",
 });
 
-// Monthly Active Users
-var monthlyActiveUsersQuery = new keenAnalysis.Query("count_unique", {
-    eventCollection: "pageviews",
-    targetProperty: "user.id",
-    timeframe: "previous_1_months",
-    timezone: "UTC"
-});
-
-// Previous 30 Days Active Users
-var thirtyDaysActiveUsersQuery = new keenAnalysis.Query("count_unique", {
-    eventCollection: "pageviews",
-    targetProperty: "user.id",
-    timeframe: "previous_30_days",
-    timezone: "UTC"
-});
-
-var dailyActiveUsersQuery = new keenAnalysis.Query("count_unique", {
-    event_collection: "pageviews",
-    target_property: "user.id",
-    timeframe: "previous_1_days",
-    timezone: "UTC"
-});
-
-var totalProjectsQuery = new keenAnalysis.Query("sum", {
+var totalProjectsQuery = new MetricParams({
     eventCollection: "node_summary",
     targetProperty: "projects.total",
     timezone: "UTC",
     timeframe: "previous_1_days",
+    respReformatter: "singleval",
 });
 
-// 7 Days back Active Users
-var weekBackThirtyDaysActiveUsersQuery = new keenAnalysis.Query("count_unique", {
-    eventCollection: "pageviews",
-    targetProperty: "user.id",
-    timeframe: getVariableDayTimeframe(7, 30),
-});
-
-// 7 Days back Active Users
-var weekBackDailyActiveUsersQuery = new keenAnalysis.Query("count_unique", {
-    eventCollection: "pageviews",
-    targetProperty: "user.id",
-    timeframe: getVariableDayTimeframe(7, 1),
-});
-
-// 28 Days back Active Users
-var monthBackThirtyDaysActiveUsersQuery = new keenAnalysis.Query("count_unique", {
-    eventCollection: "pageviews",
-    targetProperty: "user.id",
-    timeframe: getVariableDayTimeframe(28, 30),
-});
-
-// 28 Days back Active Users
-var monthBackDailyActiveUsersQuery = new keenAnalysis.Query("count_unique", {
-    eventCollection: "pageviews",
-    targetProperty: "user.id",
-    timeframe: getVariableDayTimeframe(28, 1),
-});
-
-// 364 Days back Active Users
-var yearBackThirtyDaysActiveUsersQuery = new keenAnalysis.Query("count_unique", {
-    eventCollection: "pageviews",
-    targetProperty: "user.id",
-    timeframe: getVariableDayTimeframe(364, 30),
-});
-
-// 364 Days back Active Users
-var yearBackDailyActiveUsersQuery = new keenAnalysis.Query("count_unique", {
-    eventCollection: "pageviews",
-    targetProperty: "user.id",
-    timeframe: getVariableDayTimeframe(364, 1),
-});
 
 // <+><+><+><+><+><+
 //    user data    |
@@ -417,219 +673,168 @@ var yearBackDailyActiveUsersQuery = new keenAnalysis.Query("count_unique", {
 var renderMainCounts = function() {
 
     // Active user chart!
-    var activeUserChartQuery = new keenAnalysis.Query("sum", {
+    var activeUserChartQuery = new MetricParams({
         eventCollection: "user_summary",
         interval: "daily",
-        targetProperty: "status.active",
+        targetProperty: "active",
         timeframe: "previous_800_days",
-        timezone: "UTC"
+        timezone: "UTC",
+        respReformatter: "interval",
     });
-    renderKeenMetric("#active-user-chart", "line", activeUserChartQuery, bigMetricHeight);
-
-    renderKeenMetric("#active-user-count", "metric", activeUsersQuery, bigMetricHeight);
+    renderMetric("#active-user-chart", "line", activeUserChartQuery, bigMetricHeight);
+    renderMetric("#active-user-count", "metric", activeUsersQuery, bigMetricHeight);
 
     // Daily Gain
-    var yesterday_user_count = new keenAnalysis.Query("sum", {
+    var yesterday_user_count = new MetricParams({
         eventCollection: "user_summary",
-        targetProperty: "status.active",
-        timeframe: getOneDayTimeframe(1, null)
+        targetProperty: "active",
+        timeframe: getOneDayTimeframe(1, null),
+        respReformatter: "singleval",
     });
 
-    var two_days_ago_user_count = new keenAnalysis.Query("sum", {
+    var two_days_ago_user_count = new MetricParams({
         eventCollection: "user_summary",
-        targetProperty: "status.active",
-        timeframe: getOneDayTimeframe(2, null)
+        targetProperty: "active",
+        timeframe: getOneDayTimeframe(2, null),
+        respReformatter: "singleval",
     });
-    renderCalculationBetweenTwoQueries(yesterday_user_count, two_days_ago_user_count, "#daily-user-increase", 'day', 'subtraction');
+    renderCalculationBetweenTwoQueries(yesterday_user_count, two_days_ago_user_count,
+                                       "#daily-user-increase", 'day', 'subtraction');
 
     // Monthly Gain
-    var last_month_user_count = new keenAnalysis.Query("sum", {
+    var last_month_user_count = new MetricParams({
         eventCollection: "user_summary",
-        targetProperty: "status.active",
-        timeframe: getOneDayTimeframe(null, 1)
+        targetProperty: "active",
+        timeframe: getOneDayTimeframe(null, 1),
+        respReformatter: "singleval",
     });
 
-    var two_months_ago_user_count = new keenAnalysis.Query("sum", {
+    var two_months_ago_user_count = new MetricParams({
         eventCollection: "user_summary",
-        targetProperty: "status.active",
-        timeframe: getOneDayTimeframe(null, 2)
+        targetProperty: "active",
+        timeframe: getOneDayTimeframe(null, 2),
+        respReformatter: "singleval",
     });
-    renderCalculationBetweenTwoQueries(last_month_user_count, two_months_ago_user_count, "#monthly-user-increase", 'month', 'subtraction', monthColor);
+    renderCalculationBetweenTwoQueries(last_month_user_count, two_months_ago_user_count,
+                                       "#monthly-user-increase", 'month', 'subtraction',
+                                       monthColor);
 
-    var week_ago_user_count = new keenAnalysis.Query("sum", {
+    var week_ago_user_count = new MetricParams({
         eventCollection: "user_summary",
-        targetProperty: "status.unconfirmed",
-        timeframe: getOneDayTimeframe(7, null)
+        targetProperty: "unconfirmed",
+        timeframe: getOneDayTimeframe(7, null),
+        respReformatter: "singleval",
     });
 
     // New Unconfirmed Users - # of unconfirmed users in the past 7 days
-    var yesterday_unconfirmed_user_count = new keenAnalysis.Query("sum", {
+    var yesterday_unconfirmed_user_count = new MetricParams({
         eventCollection: "user_summary",
-        targetProperty: "status.unconfirmed",
-        timeframe: getOneDayTimeframe(1, null)
+        targetProperty: "unconfirmed",
+        timeframe: getOneDayTimeframe(1, null),
+        respReformatter: "singleval",
     });
-    renderCalculationBetweenTwoQueries(yesterday_unconfirmed_user_count, week_ago_user_count, "#unverified-new-users", 'week', 'subtraction');
+    renderCalculationBetweenTwoQueries(yesterday_unconfirmed_user_count, week_ago_user_count,
+                                       "#unverified-new-users", 'week', 'subtraction');
 
 };
 
-//  Weekly User Gain metric
-var renderAverageUserGainMetric = function (results) {
+//  Weekly User Gain metrics
+var renderWeeklyUserGainMetrics = function () {
 
-    var userGainChart = new keenDataviz()
+    var weeklyUserGain = new MetricParams({
+        eventCollection: "user_summary",
+        timeframe: "previous_7_days",
+        timezone: "UTC",
+        respReformatter: "just-return",
+    });
+
+    var avgUserGainChart = new keenDataviz()
         .el("#average-gain-metric")
         .type("metric")
         .height(defaultHeight)
         .title(' ')
         .prepare();
 
-    client.run(results.queries, function (err, res) {
-        var sum = 0;
-        for (var j = 0; j < res.length - 1; j++) {
-            sum += (res[j].result - res[j + 1].result);
-        }
-        userGainChart.parseRawData({result: sum / (res.length - 1)}).render();
-    });
-
-};
-
-// User Gain Chart over past 7 days
-var renderWeeklyUserGainChart = function (results) {
-
-    var userGainChart = new keenDataviz()
+    var weeklyUserGainChart = new keenDataviz()
         .el("#user-gain-chart")
         .type("line")
         .title(' ')
         .prepare();
 
-    client.run(results.queries, function (err, res) {
-        var data = [];
-        for (var j = 0; j < res.length - 1; j++) {
-            data.push({
-                "value": res[j].result - res[j + 1].result,
-                "timeframe": results.timeframes[j]
-            });
-        }
-        userGainChart.parseRawData({result: data}).render();
-    });
-};
-
-
-// Previous 7 Days of Users by Status
-var renderPreviousWeekOfUsersByStatus = function() {
-
-    var previous_week_active_users = new keenAnalysis.Query("sum", {
-        eventCollection: "user_summary",
-        interval: "daily",
-        targetProperty: "status.active",
-        timeframe: "previous_1_week",
-        timezone: "UTC"
-    });
-
-    var previous_week_unconfirmed_users = new keenAnalysis.Query("sum", {
-        eventCollection: "user_summary",
-        interval: "daily",
-        targetProperty: "status.unconfirmed",
-        timeframe: "previous_1_week",
-        timezone: "UTC"
-    });
-
-    var previous_week_merged_users = new keenAnalysis.Query("sum", {
-        eventCollection: "user_summary",
-        interval: "daily",
-        targetProperty: "status.merged",
-        timeframe: "previous_1_week",
-        timezone: "UTC"
-    });
-
-    var previous_week_depth_users = new keenAnalysis.Query("sum", {
-        eventCollection: "user_summary",
-        interval: "daily",
-        targetProperty: "status.depth",
-        timeframe: "previous_1_week",
-        timezone: "UTC"
-    });
-
-    var previous_week_deactivated_users = new keenAnalysis.Query("sum", {
-        eventCollection: "user_summary",
-        interval: "daily",
-        targetProperty: "status.deactivated",
-        timeframe: "previous_1_week",
-        timezone: "UTC"
-    });
-
-    var chart = new keenDataviz()
+    var previousWeekOfUsersByStatusChart = new keenDataviz()
         .el("#previous-7-days-of-users-by-status")
         .height(defaultHeight)
         .type("line")
         .prepare();
 
-    client.run([
-        previous_week_active_users,
-        previous_week_unconfirmed_users,
-        previous_week_merged_users,
-        previous_week_depth_users,
-        previous_week_deactivated_users
-    ], function (err, res) {
-        var active_result = res[0].result;
-        var unconfirmed_result = res[1].result;
-        var merged_result = res[2].result;
-        var depth_result = res[3].result;
-        var deactivated_result = res[4].result;
-        var data = [];
-        var i = 0;
+    _resolveQueries([weeklyUserGain]).then(res => {
+        var newRes = reformatResponse(weeklyUserGain.params, res);
 
-        while (i < active_result.length) {
-            data[i] = {
-                timeframe: active_result[i]["timeframe"],
+        var avgDailyUserSum = 0;
+        var weeklyData = [];
+        var weekByStatusData = [];
+        var resultLen = newRes.result.length;
+
+        for (var j = 0; j < resultLen - 1; j++) {
+            var thisThing = newRes.result[j];
+            weekByStatusData[j] = {
+                timeframe: thisThing["timeframe"],
                 value: [
-                    {category: "Active", result: active_result[i].value},
-                    {category: "Unconfirmed", result: unconfirmed_result[i].value},
-                    {category: "Merged", result: merged_result[i].value},
-                    {category: "Depth", result: depth_result[i].value},
-                    {category: "Deactivated", result: deactivated_result[i].value}
+                    {category: "Active", result: thisThing.attributes.active},
+                    {category: "Unconfirmed", result: thisThing.attributes.unconfirmed},
+                    {category: "Merged", result: thisThing.attributes.merged},
+                    {category: "Deactivated", result: thisThing.attributes.deactivated},
                 ]
             };
-            if (i === active_result.length - 1) {
-                chart.parseRawData({result: data}).render();
+
+            var nextThing = newRes.result[j + 1];
+            if (nextThing) {
+                avgDailyUserSum += (thisThing.attributes.active - nextThing.attributes.active);
+                weeklyData.push({
+                    value: thisThing.attributes.active - nextThing.attributes.active,
+                    timeframe: thisThing.timeframe,
+                });
             }
-            i++;
         }
+
+        avgUserGainChart.parseRawData({result: avgDailyUserSum / (resultLen - 1)}).render();
+        weeklyUserGainChart.parseRawData({result: weeklyData}).render();
+        previousWeekOfUsersByStatusChart.parseRawData({result: weekByStatusData}).render();
     });
 };
 
-// Registrations by Email Domain
-var email_domains = new keenAnalysis.Query("count", {
-    eventCollection: "user_domain_events",
-    groupBy: "domain",
-    interval: "daily",
-    timeframe: "previous_7_days",
-    timezone: "UTC"
-});
-
 var renderEmailDomainsChart = function() {
+    // Registrations by Email Domain
+    var emailDomains = new MetricParams({
+        eventCollection: "new_user_domains",
+        groupBy: ["domain", "domain"],
+        interval: "daily",
+        timeframe: "previous_7_days",
+        timezone: "UTC",
+        respReformatter: "grouped-interval",
+    });
+
     var chart = new keenDataviz()
         .el('#user-registration-by-email-domain')
         .title(' ')
         .type('line')
         .prepare();
 
-    client.run(email_domains)
-        .then(function (res) {
-            var chartWithData = chart.data(res);
-            chartWithData.dataset.filterColumns(function (column, index) {
-                var emailThreshhold = 1;
-                for (var i = 0; i < column.length; i++) {
-                    if (column[i] > emailThreshhold) {
-                        return column;
-                    }
+    _resolveQueries([emailDomains]).then(res => {
+        var newRes = reformatResponse(emailDomains.params, res);
+        var chartWithData = chart.data(newRes);
+        chartWithData.dataset.filterColumns(function (column, index) {
+            var emailThreshhold = 1;
+            for (var i = 0; i < column.length; i++) {
+                if (column[i] > emailThreshhold) {
+                    return column;
                 }
-            });
-
-            chartWithData.render();
-        })
-        .catch(function (err) {
-            chart.message(err.message);
+            }
         });
+        chartWithData.render();
+    }).fail(function (jqXHR, textStatus, err) {
+        chart.message(err.message);
+    });
 };
 
 var NodeLogsPerUser = function() {
@@ -673,18 +878,94 @@ var NodeLogsPerUser = function() {
         });
 };
 
+var renderRawNodeMetrics = function() {
+    var propertiesAndElements = [
+        ['projects.public',               '#public-projects',              ],
+        ['projects.private',              '#private-projects',             ],
+        ['nodes.total',                   '#total-nodes',                  ],
+        ['nodes.public',                  '#public-nodes',                 ],
+        ['nodes.private',                 '#private-nodes',                ],
+        ['registered_projects.total',     '#total-registered-projects',    ],
+        ['registered_projects.public',    '#public-registered-projects',   ],
+        ['registered_projects.embargoed', '#embargoed-registered-projects',],
+        ['registered_nodes.total',        '#total-registered-nodes',       ],
+        ['registered_nodes.public',       '#public-registered-nodes',      ],
+        ['registered_nodes.embargoed',    '#embargoed-registered-nodes',   ],
+    ];
+
+    var piePropertiesAndElements = [
+        ['#total-nodes-pie',               ['nodes.public', 'nodes.private'],                              ],
+        ['#total-projects-pie',            ['projects.public', 'projects.private'],                        ],
+        ['#total-registered-nodes-pie',    ['registered_nodes.public', 'registered_nodes.embargoed'],      ],
+        ['#total-registered-projects-pie', ['registered_projects.public', 'registered_projects.embargoed'],],
+    ];
+
+    var nodeSummary = new MetricParams({
+        eventCollection: "node_summary",
+        timeframe: "previous_1_days",
+        timezone: "UTC",
+        respReformatter: "just-return",
+    });
+
+    var results = {};
+    _resolveQueries([nodeSummary]).then(res => {
+        var newRes = reformatResponse(nodeSummary.params, res);
+        var attrs = newRes.result[0].attributes;
+
+        propertiesAndElements.forEach(propAndEl => {
+            var property = propAndEl[0];
+            var element = propAndEl[1];
+
+            var chart = new keenDataviz()
+                .el(element)
+                .height(defaultHeight)
+                .title(' ')
+                .type('metric')
+                .prepare();
+
+            if (property.includes('public')) {
+                chart.colors([publicColor]);
+            }
+            else if (property.includes('private') || property.includes('embargoed')) {
+                chart.colors([privateColor]);
+            }
+
+            var finalProp = _diveIntoProperties(property, attrs);
+            chart.data({result: finalProp}).render();
+            results[property] = finalProp;
+        });
+
+        piePropertiesAndElements.forEach(piePropAndEl => {
+            var element = piePropAndEl[0];
+            var properties = piePropAndEl[1];
+
+            var publicData = properties[0];
+            var privateData = properties[1];
+
+            c3.generate({
+                bindto: element,
+                size: {height: defaultHeight*2},
+                data: {
+                    columns: [
+                        ['public', results[publicData]],
+                        ['private', results[privateData]],
+                    ],
+                    colors: {
+                        public: publicColor,
+                        private: privateColor
+                    },
+                    type : 'pie',
+                }
+            });
+        });
+    });
+};
+
 
 var UserGainMetrics = function() {
     renderMainCounts();
-
-    var weeklyUserGain = getWeeklyUserGain();
-
-    renderWeeklyUserGainChart(weeklyUserGain);
-    renderAverageUserGainMetric(weeklyUserGain);
-
+    renderWeeklyUserGainMetrics();
     renderEmailDomainsChart();
-    renderPreviousWeekOfUsersByStatus();
-
     NodeLogsPerUser();
 };
 
@@ -696,199 +977,48 @@ var UserGainMetrics = function() {
 var InstitutionMetrics = function() {
 
     // Institutional Users over past 100 Days
-    var institutional_user_chart = new keenAnalysis.Query("sum", {
+    var institutional_user_chart = new MetricParams({
         eventCollection: "institution_summary",
         interval: "daily",
         targetProperty: "users.total",
-        groupBy: "institution.name",
+        groupBy: ["institution.name", "institution_name"],
         timeframe: "previous_100_days",
-        timezone: "UTC"
+        timezone: "UTC",
+        respReformatter: "grouped-interval",
     });
-    renderKeenMetric("#institution-growth", "line", institutional_user_chart, 400);
-
+    renderMetric("#institution-growth", "line", institutional_user_chart, 400);
 
     // Total Institutional Users
-    var institutional_user_count = new keenAnalysis.Query("sum", {
+    var institutional_user_count = new MetricParams({
         eventCollection: "institution_summary",
         targetProperty: "users.total",
         timeframe: "previous_1_day",
-        timezone: "UTC"
+        timezone: "UTC",
+        respReformatter: "sum-over",  // FIXME: this should be a sum of all
     });
-    renderKeenMetric("#total-institutional-users", "metric", institutional_user_count, defaultHeight);
+    renderMetric("#total-institutional-users", "metric", institutional_user_count, defaultHeight);
 
     // Total Instutional Users / Total OSF Users
-    renderCalculationBetweenTwoQueries(institutional_user_count, activeUsersQuery, "#percentage-institutional-users", null, 'percentage');
+    renderCalculationBetweenTwoQueries(institutional_user_count, activeUsersQuery,
+                                       "#percentage-institutional-users", null, 'percentage');
 
-    // Nodes!
-
-    // Affiliated Public Nodes
-    var affiliated_public_chart = new keenAnalysis.Query("sum", {
+    // Insitutional Nodes!
+    var yesterdaysInstUserCount = new MetricParams({
         eventCollection: "institution_summary",
-        targetProperty: "nodes.public",
-        timeframe: "previous_1_days",
-        groupBy: "institution.name",
-        timezone: "UTC"
+        timeframe: "previous_1_day",
+        timezone: "UTC",
+        respReformatter: "table",
     });
-    renderKeenMetric("#affiliated-public-nodes", "table", affiliated_public_chart, institutionTableHeight);
-
-
-    // Affiliated Private Nodes
-    var affiliated_private_node_chart = new keenAnalysis.Query("sum", {
-        eventCollection: "institution_summary",
-        targetProperty: "nodes.private",
-        timeframe: "previous_1_days",
-        groupBy: "institution.name",
-        timezone: "UTC"
-    });
-    renderKeenMetric("#affiliated-private-nodes", "table", affiliated_private_node_chart, institutionTableHeight);
-
-
-    // Affiliated Public Registrations
-    var affiliated_public_registered_nodes_chart = new keenAnalysis.Query("sum", {
-        eventCollection: "institution_summary",
-        targetProperty: "registered_nodes.public",
-        timeframe: "previous_1_days",
-        groupBy: "institution.name",
-        timezone: "UTC"
-    });
-    renderKeenMetric("#affiliated-public-registered-nodes", "table", affiliated_public_registered_nodes_chart, institutionTableHeight);
-
-    // Affiliated Private Registrations
-    var affiliated_private_registered_node_chart = new keenAnalysis.Query("sum", {
-        eventCollection: "institution_summary",
-        targetProperty: "registered_nodes.embargoed_v2",
-        timeframe: "previous_1_days",
-        groupBy: "institution.name",
-        timezone: "UTC"
-    });
-    renderKeenMetric("#affiliated-embargoed-registered-nodes", "table", affiliated_private_registered_node_chart, institutionTableHeight);
-
-    // Projcets!
-
-    // Affiliated Public Projects
-    var affiliated_public_projects_chart = new keenAnalysis.Query("sum", {
-        eventCollection: "institution_summary",
-        targetProperty: "projects.public",
-        timeframe: "previous_1_days",
-        groupBy: "institution.name",
-        timezone: "UTC"
-    });
-    renderKeenMetric("#affiliated-public-projects", "table", affiliated_public_projects_chart, institutionTableHeight);
-
-
-    // Affiliated Private Projects
-    var affiliated_private_project_chart = new keenAnalysis.Query("sum", {
-        eventCollection: "institution_summary",
-        targetProperty: "projects.private",
-        timeframe: "previous_1_days",
-        groupBy: "institution.name",
-        timezone: "UTC"
-    });
-    renderKeenMetric("#affiliated-private-projects", "table", affiliated_private_project_chart, institutionTableHeight);
-
-
-    // Affiliated Public Projects Registrations
-    var affiliated_public_registered_projects_chart = new keenAnalysis.Query("sum", {
-        eventCollection: "institution_summary",
-        targetProperty: "registered_projects.public",
-        timeframe: "previous_1_days",
-        groupBy: "institution.name",
-        timezone: "UTC"
-    });
-    renderKeenMetric("#affiliated-public-registered-projects", "table", affiliated_public_registered_projects_chart, institutionTableHeight);
-
-    // Affiliated Private Projects Registrations
-    var affiliated_private_registered_projects_chart = new keenAnalysis.Query("sum", {
-        eventCollection: "institution_summary",
-        targetProperty: "registered_projects.embargoed_v2",
-        timeframe: "previous_1_days",
-        groupBy: "institution.name",
-        timezone: "UTC"
-    });
-    renderKeenMetric("#affiliated-embargoed-registered-projects", "table", affiliated_private_registered_projects_chart, institutionTableHeight);
-
-};
-
-// <+><+><+><+><+><+><+<+>+
-//   active user metrics |
-// ><+><+><+><+><+><+><+><+
-
-
-var ActiveUserMetrics = function() {
-
-    // Recent Daily Unique Sessions
-    var recentDailyUniqueSessions = new keenAnalysis.Query("count_unique", {
-        eventCollection: "pageviews",
-        targetProperty: "visitor.session",
-        interval: "daily",
-        timeframe: "previous_14_days",
-        timezone: "UTC"
-    });
-    renderKeenMetric("#recent-daily-unique-sessions", "line", recentDailyUniqueSessions, defaultHeight);
-
-    // Daily Active Users
-    var dailyActiveUsersQuery = new keenAnalysis.Query("count_unique", {
-        event_collection: "pageviews",
-        target_property: "user.id",
-        timeframe: "previous_1_days",
-        timezone: "UTC"
-
-    });
-    renderKeenMetric("#daily-active-users", "metric", dailyActiveUsersQuery, defaultHeight);
-
-    // Daily Active Users / Total Users
-    renderCalculationBetweenTwoQueries(dailyActiveUsersQuery, activeUsersQuery, "#daily-active-over-total-users", null, "percentage");
-
-
-    renderKeenMetric("#monthly-active-users", "metric", monthlyActiveUsersQuery, defaultHeight, monthColor);
-
-
-    // Monthly Active Users / Total Users
-    renderCalculationBetweenTwoQueries(monthlyActiveUsersQuery, activeUsersQuery, "#monthly-active-over-total-users", null, 'percentage', monthColor);
-
-
-    // Monthly Growth of MAU% -- Two months ago vs 1 month ago
-    var twoMonthsAgoActiveUsersQuery = new keenAnalysis.Query("count_unique", {
-        eventCollection: "pageviews",
-        targetProperty: "user.id",
-        timeframe: "previous_2_months",
-        timezone: "UTC"
-    });
-    differenceGrowthBetweenMetrics(twoMonthsAgoActiveUsersQuery, monthlyActiveUsersQuery, activeUsersQuery, "#monthly-active-user-increase", monthColor);
-
-    // Yearly Active Users
-    var yearlyActiveUsersQuery = new keenAnalysis.Query("count_unique", {
-        eventCollection: "pageviews",
-        targetProperty: "user.id",
-        timeframe: "previous_1_years",
-        timezone: "UTC"
-    });
-    renderKeenMetric("#yearly-active-users", "metric", yearlyActiveUsersQuery, defaultHeight, yearColor);
-
-    // Yearly Active Users / Total Users
-    renderCalculationBetweenTwoQueries(yearlyActiveUsersQuery, activeUsersQuery, "#yearly-active-over-total-users", null, 'percentage', yearColor);
-
-    // Average Projects per User
-    renderCalculationBetweenTwoQueries(totalProjectsQuery, activeUsersQuery, "#projects-per-user", null, 'division');
-
-    // Average Projects per MAU
-    renderCalculationBetweenTwoQueries(totalProjectsQuery, monthlyActiveUsersQuery, "#projects-per-monthly-user", null, 'division');
-};
-
-// <+><+><+><+><+><+><+<+>+
-//   healthy user metrics |
-// ><+><+><+><+><+><+><+><+
-
-var HealthyUserMetrics = function() {
-
-    // stickiness ratio - DAU/MAU
-    renderCalculationBetweenTwoQueries(dailyActiveUsersQuery, thirtyDaysActiveUsersQuery, "#stickiness-ratio-1-day-ago", null, "percentage");
-     // stickiness ratio - DAU/MAU for 1 week ago
-    renderCalculationBetweenTwoQueries(weekBackDailyActiveUsersQuery, weekBackThirtyDaysActiveUsersQuery , "#stickiness-ratio-1-week-ago", null, "percentage");
-    // stickiness ratio - DAU/MAU for 4 weeks ago
-    renderCalculationBetweenTwoQueries(monthBackDailyActiveUsersQuery, monthBackThirtyDaysActiveUsersQuery , "#stickiness-ratio-4-weeks-ago", null, "percentage");
-    // stickiness ratio - DAU/MAU for 52 weeks ago
-    renderCalculationBetweenTwoQueries(yearBackDailyActiveUsersQuery, yearBackThirtyDaysActiveUsersQuery , "#stickiness-ratio-52-weeks-ago", null, "percentage");
+    renderMetricsForInsts(yesterdaysInstUserCount, [
+        ["#affiliated-public-nodes",                  "nodes.public",                    ],
+        ["#affiliated-private-nodes",                 "nodes.private",                   ],
+        ["#affiliated-public-registered-nodes",       "registered_nodes.public",         ],
+        ["#affiliated-embargoed-registered-nodes",    "registered_nodes.embargoed_v2",   ],
+        ["#affiliated-public-projects",               "projects.public",                 ],
+        ["#affiliated-private-projects",              "projects.private",                ],
+        ["#affiliated-public-registered-projects",    "registered_projects.public",      ],
+        ["#affiliated-embargoed-registered-projects", "registered_projects.embargoed_v2",],
+    ]);
 };
 
 
@@ -897,86 +1027,10 @@ var HealthyUserMetrics = function() {
 // ><+><+><+><><+><+
 
 var RawNumberMetrics = function() {
-
-    renderKeenMetric("#total-projects", "metric", totalProjectsQuery, defaultHeight);
-
-    var propertiesAndElements = {
-        'projects.public': '#public-projects',
-        'projects.private': '#private-projects',
-        'nodes.total': '#total-nodes',
-        'nodes.public': '#public-nodes',
-        'nodes.private': '#private-nodes',
-        'registered_projects.total': '#total-registered-projects',
-        'registered_projects.public': '#public-registered-projects',
-        'registered_projects.embargoed': '#embargoed-registered-projects',
-        'registered_nodes.total': '#total-registered-nodes',
-        'registered_nodes.public': '#public-registered-nodes',
-        'registered_nodes.embargoed': '#embargoed-registered-nodes'
-    };
-
-    var piePropertiesAndElements = {
-        '#total-nodes-pie': ['nodes.public', 'nodes.private'],
-        '#total-projects-pie': ['projects.public', 'projects.private'],
-        '#total-registered-nodes-pie': ['registered_nodes.public', 'registered_nodes.embargoed'],
-        '#total-registered-projects-pie': ['registered_projects.public', 'registered_projects.embargoed']
-    };
-
-    var graphPromises = [];
-    for (var key in propertiesAndElements) {
-        if (propertiesAndElements.hasOwnProperty(key)) {
-            graphPromises.push(client.query('sum', {
-                event_collection: "node_summary",
-                target_property: key,
-                timeframe: "previous_1_days",
-                timezone: "UTC"
-            }))
-        }
-    }
-
-    var results = {};
-    Promise.all(graphPromises).then(values => {
-        for (var i=0; i<values.length; i++) {
-            var chart = new keenDataviz()
-                .el(propertiesAndElements[values[i].query.target_property])
-                .height(defaultHeight)
-                .title(' ')
-                .type('metric')
-                .prepare();
-
-            if (values[i].query.target_property.includes('public')) {
-                chart.colors([publicColor]);
-            } else if (values[i].query.target_property.includes('private') || values[i].query.target_property.includes('embargoed')) {
-                chart.colors([privateColor]);
-            }
-            chart.data(values[i]).render();
-
-            var targetProperty = values[i].query.target_property.toString();
-            results[targetProperty] = values[i].result;
-        }
-
-        for (var element in piePropertiesAndElements) {
-            if (piePropertiesAndElements.hasOwnProperty(element)) {
-                var publicData = piePropertiesAndElements[element][0];
-                var privateData = piePropertiesAndElements[element][1];
-                c3.generate({
-                    bindto: element,
-                    size: {height: defaultHeight*2},
-                    data: {
-                        columns: [
-                            ['public', results[publicData]],
-                            ['private', results[privateData]],
-                        ],
-                        colors: {
-                            public: publicColor,
-                            private: privateColor
-                        },
-                        type : 'pie',
-                    }
-                });
-            }
-        }
-    });
+    renderMetric("#total-projects", "metric", totalProjectsQuery, defaultHeight);
+    renderRawNodeMetrics();
 };
+
 
 // <+><+><+><><>+
 //     addons   |
@@ -984,15 +1038,17 @@ var RawNumberMetrics = function() {
 
 var AddonMetrics = function() {
     // Previous 7 days of linked addon by addon name
-    var linked_addon = new keenAnalysis.Query("sum", {
-        eventCollection: "addon_snapshot",
-        targetProperty: "users.linked",
-        groupBy: ["provider.name"],
+    var linked_addon = new MetricParams({
+        eventCollection: "storage_addon_usage",
         interval: "daily",
         timeframe: "previous_8_days",
-        timezone: "UTC"
+        timezone: "UTC",
+        listKey: "usage_by_addon",
+        nameKey: "addon_shortname",
+        resultKey: "linked_usersettings.total",
+        respReformatter: "extract-list"
     });
-    renderKeenMetric('#previous-7-days-of-linked-addon-by-addon-name', "line", linked_addon, defaultHeight);
+    renderMetric('#previous-7-days-of-linked-addon-by-addon-name', "line", linked_addon, defaultHeight);
 };
 
 
@@ -1001,25 +1057,35 @@ var AddonMetrics = function() {
 // ><+><+><+<+><+
 
 var PreprintMetrics = function() {
-    var preprint_created = new keenAnalysis.Query("sum", {
-        eventCollection: "preprint_summary",
-        targetProperty: "provider.total",
-        groupBy: ["provider.name"],
-        interval: "daily",
-        timeframe: "previous_30_days",
-        timezone: "UTC"
-    });
-
-    renderKeenMetric("#preprints-added", "line", preprint_created, defaultHeight);
+    renderPreprintMetrics();
 };
 
+function renderPreprintMetrics(timeframe) {
+    if (timeframe === undefined) {
+        timeframe = "previous_30_days";
+    }
 
-//≠≠≠≠≠≠≠≠≠≠≠≠≠≠≠≠≠≠≠≠≠≠≠≠≠≠≠
+    var preprint_created = new MetricParams({
+        eventCollection: "preprint_summary",
+        targetProperty: "preprint_count",
+        groupBy: ["provider.name", "provider_key"],
+        interval: "daily",
+        timeframe: timeframe,
+        timezone: "UTC",
+        respReformatter: "grouped-interval",
+    });
+
+    renderMetric("#preprints-added", "line", preprint_created, defaultHeight);
+}
+
+
+// <+><+><+><><>+
 //   file download counts   |
-//≠≠≠≠≠≠≠≠≠≠≠≠≠≠≠≠≠≠≠≠≠≠≠≠≠≠≠
+// ><+><+><+<+><+
 
 var DownloadMetrics = function() {
 
+    // ********** legacy keen **********
     var totalDownloadsQuery = new keenAnalysis.Query("count", {
         eventCollection: "file_stats",
         timeframe: 'previous_1_days',
@@ -1030,8 +1096,10 @@ var DownloadMetrics = function() {
             timezone: "UTC"
         }]
     });
-    renderKeenMetric("#number-of-downloads", "metric", totalDownloadsQuery, defaultHeight, defaultColor, publicClient);
+    renderKeenMetric("#number-of-downloads", "metric", totalDownloadsQuery,
+                     defaultHeight, defaultColor, publicClient);
 
+    // ********** legacy keen **********
     var uniqueDownloadsQuery = new keenAnalysis.Query("count_unique", {
         eventCollection: "file_stats",
         timeframe: 'previous_1_days',
@@ -1043,30 +1111,40 @@ var DownloadMetrics = function() {
             timezone: "UTC"
         }]
     });
-    renderKeenMetric("#number-of-unique-downloads", "metric", uniqueDownloadsQuery, defaultHeight, defaultColor, publicClient);
+    renderKeenMetric("#number-of-unique-downloads", "metric", uniqueDownloadsQuery,
+                     defaultHeight, defaultColor, publicClient);
 
-    var downloadCount = new keenAnalysis.Query("sum", {
-        eventCollection: "download_count_summary",
-        interval: "daily",
-        targetProperty: "files.total",
-        timeframe: "previous_30_days",
-        timezone: "UTC"
-    });
-
-    renderKeenMetric("#download-counts", "line", downloadCount, defaultHeight);
+    renderDownloadMetrics();
 };
+
+function renderDownloadMetrics(timeframe) {
+    if (timeframe === undefined) {
+        timeframe = "previous_30_days";
+    }
+
+    var downloadCount = new MetricParams({
+        eventCollection: "download_count",
+        interval: "daily",
+        targetProperty: "daily_file_downloads",
+        timeframe: timeframe,
+        timezone: "UTC",
+        respReformatter: "interval",
+    });
+    renderMetric("#download-counts", "line", downloadCount, defaultHeight);
+}
+
+
 
 
 module.exports = {
     UserGainMetrics: UserGainMetrics,
     NodeLogsPerUser: NodeLogsPerUser,
     InstitutionMetrics: InstitutionMetrics,
-    ActiveUserMetrics: ActiveUserMetrics,
-    HealthyUserMetrics:HealthyUserMetrics,
     RawNumberMetrics: RawNumberMetrics,
     AddonMetrics: AddonMetrics,
     PreprintMetrics: PreprintMetrics,
     DownloadMetrics: DownloadMetrics,
-    KeenRenderMetrics: renderKeenMetric
-
+    KeenRenderMetrics: renderKeenMetric,
+    RenderPreprintMetrics: renderPreprintMetrics,
+    RenderDownloadMetrics: renderDownloadMetrics,
 };
