+++ admin/static/js/pages/metrics-page.js
@@ -18,14 +18,6 @@ keenAnalysis.ready(function() {
         Metrics.InstitutionMetrics();
     };
 
-    $('#active-user-tab')[0].onclick = function() {
-        Metrics.ActiveUserMetrics();
-    };
-
-    $('#healthy-user-tab')[0].onclick = function() {
-        Metrics.HealthyUserMetrics();
-    };
-
     $('#raw-numbers-tab')[0].onclick = function() {
         Metrics.RawNumberMetrics();
     };
@@ -39,33 +31,16 @@ keenAnalysis.ready(function() {
         Metrics.DownloadMetrics();
     };
     $('#preprints-range')[0].onclick = function() {
-        var preprint_created = new keenAnalysis.Query('sum', {
-            eventCollection: 'preprint_summary',
-            targetProperty: 'provider.total',
-            groupBy: ['provider.name'],
-            interval: 'daily',
-            timeframe: {
-                'start': $('#start-date')[0].value,
-                'end': $('#end-date')[0].value
-            },
-            timezone: "UTC"
+        Metrics.RenderPreprintMetrics({
+            start: $('#start-date')[0].value,
+            end: $('#end-date')[0].value
         });
-
-        Metrics.KeenRenderMetrics("#preprints-added", "line", preprint_created, 200);
     };
     $('#downloads-range')[0].onclick = function() {
-        var downloadCount = new keenAnalysis.Query('sum', {
-            eventCollection: 'download_count_summary',
-            targetProperty: 'files.total',
-            interval: 'daily',
-            timeframe: {
-                'start': $('#start-date-downloads')[0].value,
-                'end': $('#end-date-downloads')[0].value
-            },
-            timezone: "UTC"
+        Metrics.RenderDownloadMetrics({
+            start: $('#start-date-downloads')[0].value,
+            end: $('#end-date-downloads')[0].value
         });
-
-        Metrics.KeenRenderMetrics("#download-counts", "line", downloadCount, 200);
     };
 
 });
