+++ admin/webpack.admin.config.js
@@ -34,7 +34,6 @@ var config = Object.assign({}, common, {
         'admin-registration-edit-page': staticAdminPath('js/pages/admin-registration-edit-page.js'),
         'dashboard': staticAdminPath('js/sales_analytics/dashboard.js'),
         'metrics-page': staticAdminPath('js/pages/metrics-page.js'),
-        'metrics-page-mw': staticAdminPath('js/pages/metrics-page-mw.js'),
         'banners': staticAdminPath('js/banners/banners.js'),
         'brands': staticAdminPath('js/brands/brands.js'),
         'maintenance': staticAdminPath('js/maintenance/maintenance.js'),
